PMM-4534 Added initial TLS support for PostgreSQL
It is initial support and not full support because in this step we added
just 2 fields to signal ssl enabled= on/off and ssl mode =
require/verify-full but in order to fully support verify-full wee need
to handle certificates.
Even that, verify-full should work since the PostgreSQL can handle
certificates in the DSN used when started the exporter.