PMM-5026 Added test for tables count (#66)
* PMM-5026 Added test for tables count

Updated the MySQLService Add method to update the tables count after the
connection checker is being called.

* PMM-5026 Skip MySQL Add test if no agents

If there are no connected agents, let's skip the test instead of making
it fail. This is because not in all testing environments we will have
connected pmm agents.

* PMM-5026 cosmetic change

* PMM-5026 Skipping test

Made test to skip until we have the new environment were we can trust
there will be connected agents to run all tests.