SAAS-487 Update PlatformSignUp flow
This commit updates the PlatformSignUp API to use use just email,
firstName and lastName to register the user. Password is no longer
set using the API but by the user using the validation email.

We also remove platform_test.go since they are not unit tests and
interact with Auth service which we already test for in API tests.