PMM-9614 Add new migration for MongoDB options
MongoDB agent configuration is stored in Postgres database
as a JSON blob. Usually, when we add new fields to the
configuration, JSON blob can be used in a backward
compatible manner. However, we ended up updating usage of
stats_collections field from a string to string array. This
requires a new migration, to allow newer version of server
to work with data stored by older server versions.