Updated cluster name generation
During QA, Beata noticed that if you provide the image name but not the
cluster name, the version used to generate the cluster name was taken
from the recommended version and not from the provided image path.
Now the version used to generate the name is being taken from the
component image path.